requirements.txt
python-dotenv==1.0.0
setuptools==65.5.0
spotipy==2.23.0
